fix isRouteSupported and separate constants
