+++ api/_bridges/cctp/strategy.ts
@@ -10,139 +10,19 @@ import { CrossSwap, CrossSwapQuotes } from "../../_dexes/types";
 import { AppFee, CROSS_SWAP_TYPE } from "../../_dexes/utils";
 import { Token } from "../../_dexes/types";
 import { InvalidParamError } from "../../_errors";
-import { CHAIN_IDs, TOKEN_SYMBOLS_MAP, CHAINS } from "../../_constants";
 import { ConvertDecimals } from "../../_utils";
 import { tagIntegratorId, tagSwapApiMarker } from "../../_integrator-id";
-import { toBytes32 } from "../../_address";
+import {
+  CCTP_SUPPORTED_CHAINS,
+  CCTP_SUPPORTED_TOKENS,
+  CCTP_FINALITY_THRESHOLDS,
+  CCTP_FILL_TIME_ESTIMATES,
+  getCctpTokenMessengerAddress,
+  getCctpDomainId,
+  encodeDepositForBurn,
+} from "./utils/constants";
 
 const name = "cctp";
-
-const CCTP_SUPPORTED_CHAINS = [
-  CHAIN_IDs.MAINNET,
-  CHAIN_IDs.ARBITRUM,
-  CHAIN_IDs.BASE,
-  CHAIN_IDs.HYPEREVM,
-  CHAIN_IDs.INK,
-  CHAIN_IDs.OPTIMISM,
-  CHAIN_IDs.POLYGON,
-  CHAIN_IDs.SOLANA,
-  CHAIN_IDs.UNICHAIN,
-  CHAIN_IDs.WORLD_CHAIN,
-];
-
-const CCTP_SUPPORTED_TOKENS = [TOKEN_SYMBOLS_MAP.USDC];
-
-const CCTP_FINALITY_THRESHOLDS = {
-  fast: 1000,
-  standard: 2000,
-};
-
-// CCTP TokenMessenger contract addresses
-const DEFAULT_CCTP_TOKEN_MESSENGER_ADDRESS =
-  "0x28b5a0e9C621a5BadaA536219b3a228C8168cf5d";
-
-const CCTP_TOKEN_MESSENGER_ADDRESS_OVERRIDES: Record<number, string> = {
-  [CHAIN_IDs.SOLANA]: "CCTPV2vPZJS2u2BBsUoscuikbYjnpFmbFsvVuJdgUMQe",
-};
-
-const getCctpTokenMessengerAddress = (chainId: number): string => {
-  return (
-    CCTP_TOKEN_MESSENGER_ADDRESS_OVERRIDES[chainId] ||
-    DEFAULT_CCTP_TOKEN_MESSENGER_ADDRESS
-  );
-};
-
-const getCctpDomainId = (chainId: number): number => {
-  const chainInfo = CHAINS[chainId];
-  if (!chainInfo || typeof chainInfo.cctpDomain !== "number") {
-    throw new InvalidParamError({
-      message: `CCTP domain not found for chain ID ${chainId}`,
-    });
-  }
-  return chainInfo.cctpDomain;
-};
-
-// CCTP TokenMessenger depositForBurn ABI
-const CCTP_DEPOSIT_FOR_BURN_ABI = {
-  inputs: [
-    {
-      internalType: "uint256",
-      name: "amount",
-      type: "uint256",
-    },
-    {
-      internalType: "uint32",
-      name: "destinationDomain",
-      type: "uint32",
-    },
-    {
-      internalType: "bytes32",
-      name: "mintRecipient",
-      type: "bytes32",
-    },
-    {
-      internalType: "address",
-      name: "burnToken",
-      type: "address",
-    },
-    {
-      internalType: "bytes32",
-      name: "destinationCaller",
-      type: "bytes32",
-    },
-    {
-      internalType: "uint256",
-      name: "maxFee",
-      type: "uint256",
-    },
-    {
-      internalType: "uint32",
-      name: "minFinalityThreshold",
-      type: "uint32",
-    },
-  ],
-  name: "depositForBurn",
-  outputs: [],
-  stateMutability: "nonpayable",
-  type: "function",
-};
-
-const encodeDepositForBurn = (params: {
-  amount: BigNumber;
-  destinationDomain: number;
-  mintRecipient: string;
-  burnToken: string;
-  destinationCaller: string;
-  maxFee: BigNumber; // Required, use BigNumber.from(0) for standard transfer
-  minFinalityThreshold: number; // use 2000 for standard transfer
-}): string => {
-  const iface = new ethers.utils.Interface([CCTP_DEPOSIT_FOR_BURN_ABI]);
-
-  return iface.encodeFunctionData("depositForBurn", [
-    params.amount,
-    params.destinationDomain,
-    toBytes32(params.mintRecipient),
-    params.burnToken,
-    toBytes32(params.destinationCaller),
-    params.maxFee,
-    params.minFinalityThreshold,
-  ]);
-};
-
-// CCTP estimated fill times in seconds
-// Soruce: https://developers.circle.com/cctp/required-block-confirmations
-const CCTP_FILL_TIME_ESTIMATES: Record<number, number> = {
-  [CHAIN_IDs.MAINNET]: 19 * 60,
-  [CHAIN_IDs.ARBITRUM]: 19 * 60,
-  [CHAIN_IDs.BASE]: 19 * 60,
-  [CHAIN_IDs.HYPEREVM]: 5,
-  [CHAIN_IDs.INK]: 30 * 60,
-  [CHAIN_IDs.OPTIMISM]: 19 * 60,
-  [CHAIN_IDs.POLYGON]: 8,
-  [CHAIN_IDs.SOLANA]: 25,
-  [CHAIN_IDs.UNICHAIN]: 19 * 60,
-  [CHAIN_IDs.WORLD_CHAIN]: 19 * 60,
-};
 
 const capabilities: BridgeCapabilities = {
   ecosystems: ["evm", "svm"],
@@ -158,7 +38,7 @@ const capabilities: BridgeCapabilities = {
 
 /**
  * CCTP (Cross-Chain Transfer Protocol) bridge strategy for native USDC transfers.
- * Supports Circle's CCTP for burning USDC on source chain and minting on destination chain.
+ * Supports Circle's CCTP for burning USDC on source chain.
  */
 export function getCctpBridgeStrategy(): BridgeStrategy {
   const getEstimatedFillTime = (originChainId: number): number => {
@@ -198,6 +78,8 @@ export function getCctpBridgeStrategy(): BridgeStrategy {
     if (!isOriginChainSupported || !isDestinationChainSupported) {
       return false;
     }
+
+    return true;
   };
 
   const assertSupportedRoute = (params: {
